+++ packages/create-apexjs/templates/default/server/api/hello.ts
@@ -1,4 +1,4 @@
-import { defineApexRoute } from 'apexjs-core'
+import { defineApexRoute } from '@apex-stack/core'
 import { z } from 'zod'
 
 export default defineApexRoute({